refactor: Introduce separate config package
