+++ cmd/relayer_exporter/relayer_exporter.go
@@ -7,7 +7,7 @@ import (
 	"os"
 
 	"github.com/archway-network/relayer_exporter/pkg/collector"
-	"github.com/archway-network/relayer_exporter/pkg/ibc"
+	"github.com/archway-network/relayer_exporter/pkg/config"
 	log "github.com/archway-network/relayer_exporter/pkg/logger"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
@@ -37,21 +37,24 @@ func main() {
 		os.Exit(0)
 	}
 
-	cfg, err := ibc.NewConfig(*configPath)
+	cfg, err := config.NewConfig(*configPath)
 	if err != nil {
 		log.Fatal(err.Error())
 	}
 
 	log.Info(fmt.Sprintf("Getting IBC paths from %s/%s/%s on GitHub", cfg.GitHub.Org, cfg.GitHub.Repo, cfg.GitHub.IBCDir))
 
+	// TODO: Add a feature to refresh paths at configured interval
 	paths, err := cfg.IBCPaths()
 	if err != nil {
 		log.Fatal(err.Error())
 	}
 
+	rpcs := cfg.GetRPCsMap()
+
 	clientsCollector := collector.IBCClientsCollector{
-		Config: *cfg,
-		Paths:  paths,
+		RPCs:  rpcs,
+		Paths: paths,
 	}
 
 	prometheus.MustRegister(clientsCollector)
