+++ pkg/collector/collector.go
@@ -3,7 +3,7 @@ package collector
 import (
 	"github.com/archway-network/relayer_exporter/pkg/ibc"
 	log "github.com/archway-network/relayer_exporter/pkg/logger"
-	gorelayer "github.com/cosmos/relayer/v2/relayer"
+	"github.com/cosmos/relayer/v2/relayer"
 	"github.com/prometheus/client_golang/prometheus"
 	"go.uber.org/zap"
 )
@@ -21,8 +21,8 @@ var clientExpiry = prometheus.NewDesc(
 )
 
 type IBCClientsCollector struct {
-	Config ibc.Config
-	Paths  []*gorelayer.IBCdata
+	RPCs  map[string]string
+	Paths []*relayer.IBCdata
 }
 
 func (cc IBCClientsCollector) Describe(_ chan<- *prometheus.Desc) {}
@@ -30,7 +30,7 @@ func (cc IBCClientsCollector) Describe(_ chan<- *prometheus.Desc) {}
 func (cc IBCClientsCollector) Collect(ch chan<- prometheus.Metric) {
 	log.Debug("Start collecting", zap.String("metric", clientExpiryMetricName))
 
-	clients := cc.Config.GetClientsInfos(cc.Paths)
+	clients := ibc.GetClientsInfos(cc.Paths, cc.RPCs)
 
 	for _, c := range clients {
 		path := c.PathName()
