+++ pkg/ibc/ibc.go
@@ -2,22 +2,16 @@ package ibc
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"os"
-	"strings"
 	"time"
 
 	log "github.com/archway-network/relayer_exporter/pkg/logger"
 	"github.com/cosmos/relayer/v2/relayer"
 	"github.com/cosmos/relayer/v2/relayer/chains/cosmos"
 	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-github/v54/github"
-	"gopkg.in/yaml.v3"
 )
 
 const (
-	ibcPathSuffix  = ".json"
 	rpcTimeout     = "10s"
 	keyringBackend = "test"
 )
@@ -51,31 +45,7 @@ func (ci ClientsInfo) PathName() string {
 	return fmt.Sprintf("%s<->%s", chainAID, chainBID)
 }
 
-type RPC struct {
-	ChainID string `yaml:"chainId"`
-	URL     string `yaml:"url"`
-}
-
-type Config struct {
-	RPCs   []RPC `yaml:"rpc"`
-	GitHub struct {
-		Org    string `yaml:"org"`
-		Repo   string `yaml:"repo"`
-		IBCDir string `yaml:"dir"`
-	} `yaml:"github"`
-}
-
-func (c *Config) getRPCs() map[string]string {
-	rpcs := map[string]string{}
-
-	for _, rpc := range c.RPCs {
-		rpcs[rpc.ChainID] = rpc.URL
-	}
-
-	return rpcs
-}
-
-func (c *Config) GetClientsInfos(ibcs []*relayer.IBCdata) []ClientsInfo {
+func GetClientsInfos(ibcs []*relayer.IBCdata, rpcs map[string]string) []ClientsInfo {
 	num := len(ibcs)
 
 	out := make(chan ClientsInfo, num)
@@ -83,7 +53,7 @@ func (c *Config) GetClientsInfos(ibcs []*relayer.IBCdata) []ClientsInfo {
 
 	for i := 0; i < num; i++ {
 		go func(i int) {
-			clientsInfo, err := c.GetClientsInfo(ibcs[i])
+			clientsInfo, err := GetClientsInfo(ibcs[i], rpcs)
 			if err != nil {
 				out <- ClientsInfo{}
 
@@ -107,10 +77,8 @@ func (c *Config) GetClientsInfos(ibcs []*relayer.IBCdata) []ClientsInfo {
 	return clientsInfos
 }
 
-func (c *Config) GetClientsInfo(ibc *relayer.IBCdata) (ClientsInfo, error) {
+func GetClientsInfo(ibc *relayer.IBCdata, rpcs map[string]string) (ClientsInfo, error) {
 	clientsInfo := ClientsInfo{}
-
-	rpcs := c.getRPCs()
 
 	cdA := ChainData{
 		ChainID:  ibc.Chain1.ChainName,
@@ -151,60 +119,6 @@ func (c *Config) GetClientsInfo(ibc *relayer.IBCdata) (ClientsInfo, error) {
 	}
 
 	return clientsInfo, nil
-}
-
-func (c *Config) IBCPaths() ([]*relayer.IBCdata, error) {
-	ctx := context.Background()
-
-	client := github.NewClient(nil)
-
-	_, ibcDir, _, err := client.Repositories.GetContents(ctx, c.GitHub.Org, c.GitHub.Repo, c.GitHub.IBCDir, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	ibcs := []*relayer.IBCdata{}
-
-	for _, file := range ibcDir {
-		if strings.HasSuffix(*file.Path, ibcPathSuffix) {
-			content, _, _, err := client.Repositories.GetContents(ctx, c.GitHub.Org, c.GitHub.Repo, *file.Path, nil)
-			if err != nil {
-				return nil, err
-			}
-
-			ibc := &relayer.IBCdata{}
-
-			c, err := content.GetContent()
-			if err != nil {
-				return nil, err
-			}
-
-			if err = json.Unmarshal([]byte(c), &ibc); err != nil {
-				return nil, err
-			}
-
-			ibcs = append(ibcs, ibc)
-		}
-	}
-
-	return ibcs, nil
-}
-
-func NewConfig(configPath string) (*Config, error) {
-	config := &Config{}
-
-	file, err := os.Open(configPath)
-	if err != nil {
-		return nil, err
-	}
-	defer file.Close()
-
-	d := yaml.NewDecoder(file)
-	if err := d.Decode(&config); err != nil {
-		return nil, err
-	}
-
-	return config, nil
 }
 
 func prepChain(cd ChainData) (*relayer.Chain, error) {
